Issue.1011
Replace for-based search by STL algoritms
Move real_room()/real_mobile() to utils
