+++ src/magic/magic_rooms.cpp
@@ -4,6 +4,7 @@
 #include "modify.h"
 #include "chars/char.h"
 #include "magic.h" //Включено ради material_component_processing, который надо, по-хлорошему, вообще в отдеьный модуль.
+#include "utils/id_converter.h"
 
 #include <iomanip>
