+++ src/magic/spells.cpp
@@ -15,6 +15,7 @@
 #include "magic/spells.h"
 
 #include "utils/utils_char_obj.inl"
+#include "utils/id_converter.h"
 #include "chars/world.characters.h"
 #include "cmd/follow.h"
 #include "cmd/hire.h"
