+++ src/utils/id_converter.cpp
@@ -1,7 +1,6 @@
 #include "id_converter.h"
 
-#include "world_objects.h"
-#include "zone.table.h"
+#include "global_objects.h"
 
 zone_rnum get_zone_rnum_by_room_vnum(room_vnum vnum) {
 	for (zone_rnum counter = 0; counter < static_cast<zone_rnum>(zone_table.size()); counter++)
@@ -23,6 +22,38 @@ zone_rnum get_zone_rnum_by_mob_vnum(mob_vnum vnum) {
 
 zone_rnum get_zone_rnum_by_zone_vnum(zone_vnum zone) {
 	return get_zone_rnum_by_room_vnum(zone * 100);
+}
+
+// returns the real number of the room with given virtual number
+room_rnum real_room(room_vnum vnum) {
+	const auto it_room = std::lower_bound(world.begin(), world.end(), vnum, [](const ROOM_DATA *room, room_vnum value) {
+		return room->number < value;
+	});
+
+	return (it_room == world.end() || (*it_room)->number != vnum) ?
+		NOWHERE : std::distance(world.begin(), it_room);
+}
+
+// returns the real number of the monster with given virtual number
+mob_rnum real_mobile(mob_vnum vnum) {
+	mob_rnum bot, top, mid;
+
+	bot = 0;
+	top = top_of_mobt;
+
+	// perform binary search on mob-table
+	for (;;) {
+		mid = (bot + top) / 2;
+
+		if ((mob_index + mid)->vnum == vnum)
+			return (mid);
+		if (bot >= top)
+			return (-1);
+		if ((mob_index + mid)->vnum > vnum)
+			top = mid - 1;
+		else
+			bot = mid + 1;
+	}
 }
 
 // vim: ts=4 sw=4 tw=0 noet syntax=cpp :
