+++ src/utils/utils_find_obj_id_by_vnum.cpp
@@ -3,6 +3,7 @@
 #include "dg_script/dg_scripts.h"
 #include "world_objects.h"
 #include "chars/char.h"
+#include "utils/id_converter.h"
 
 class TriggerLookup {
  public:
